Packets refactoring, part III  - Restore pakets init and initFromByteArray  - Add sensors (IMU and Feet) read  - Add sensors offset to config file and set this offsets in    packet initialization  - Replace offsets, limits in config to double  - Moar small fixes

+++ src/AR60xHWDriver/HardwareInterfaces/IAR60xPowerControl.h
@@ -4,7 +4,7 @@
 class IAR60xPowerControl
 {
 public:
-    virtual PowerState::PowerSupplyState JointGetSupplyState(int joint) = 0;
+    virtual PowerState::PowerSupplyState JointGetSupplyState(uint8_t joint) = 0;
     virtual PowerState::PowerSupplyState PowerGetSupplyState(PowerData::PowerSupplies supply) = 0;
     virtual void SupplySetOnOff(PowerData::PowerSupplies supply, bool onOffState) = 0;
     virtual bool SupplyGetOnOff(PowerData::PowerSupplies supply) = 0;
